tests/cocotb/pyproject.toml
# Copyright (c) 2025-2026 Antmicro <www.antmicro.com>
# SPDX-License-Identifier: Apache-2.0

[project]
name = "guineveer-cocotb-tests"
version = "0.0.1"
dependencies = [
  "cocotb==1.9.2",
  "cocotbext-ahb==0.4.9",
  "cocotbext-uart==0.1.2",
  "cocotbext-i3c @ git+https://github.com/antmicro/cocotbext-i3c@dfe72c4fb3dee668011432046b961b8c488027e9",
  "junitparser==3.2.0"
]

[tool.setuptools]
packages = []

[tool.ruff]
target-version = "py38"
line-length = 100
extend-exclude = [
  "__pycache__",
]

[tool.ruff.lint]
select = [
    "A",   # flake8-builtins
    "ARG", # flake8-unused-arguments
    "B",   # flake8-bugbear
    "E",   # pycodestyle
    "EM",  # flake8-errmsg
    "F",   # Pyflakes
    "I",   # isort
    "ICN", # flake8-import-conventions
    "ISC", # flake8-implicit-str-concat
    "N",   # pep8-naming
    "Q",   # flake8-quotes
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
    "W",   # pycodestyle
    "YTT", # flake8-2020
]
ignore = [
    "B011",   # Do not `assert False` #FIXME
    "E203",   # Whitespace before ':'
    "D105",   # Missing docstring in magic method
    "D204",   # 1 blank line required after class docstring
    "D205",   # 1 blank line required between summary line and description
    "D200",   # One-line docstring should fit on one line
    "ISC001", # Single-line-implicit-string-concatenation
    "EM101",  # Checks for the use of string literals in exception constructors
    "EM102",  # Checks for the use of f-strings in exception constructors
    "N802",   # Checks for functions names that do not follow the `snake_case` naming convention
    "N805"    # Checks for instance methods that use a name other than `self` for their first argument
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
